Funcionalidades com relação ao usuario
Funcionalidades com relação ao usuario terminadas

Buscar todos os usuários
Buscar um usuario
Criar um usuario
Apagar um usuario
Atualizar um usuario

+++ src/models/user.rs
@@ -1,17 +1,17 @@
+use serde::{Deserialize, Serialize};
 
+#[derive(Deserialize, Serialize)]
 pub struct User {
-    id: i32,
+    id: i64,
     name: String,
     password: String
 }
 
-
-
 impl User {
-    pub fn new (id: i32, name: String, password: String) -> User {
+    pub fn new(id: i64, name: String, password: String) -> User {
         User { id, name, password }
     }
-    pub fn get_id(&self) -> i32 {
+    pub fn get_id(&self) -> i64 {
         self.id.clone()
     }
     pub fn get_name(&self) -> String {
